Remove ability to change the LOB root of a license contract
The root contract has functions that operate all functions in the
license contract that can only be done by the LOB root and thus there is
no point in allowing the root to change since the owner of the LOB root
can already change.

Also, this will allow the certificate text to be stored in the LOB root.